+++ src/inspect_ai/model/_model.py
@@ -46,6 +46,7 @@ from ._generate_config import (
 )
 from ._model_call import ModelCall
 from ._model_output import ModelOutput, ModelUsage
+from ._trace import trace_assistant_message
 
 logger = logging.getLogger(__name__)
 
@@ -157,16 +158,6 @@ class ModelAPI(abc.ABC):
 
     def tools_required(self) -> bool:
         """Any tool use in a message stream means that tools must be passed."""
-        return False
-
-    def provides_event(self) -> bool:
-        """Indicates that this model provider yields a ModelEvent.
-
-        This is done so the code that wraps ModelAPI calls knows that it
-        doesn't need to record a ModelEvent. A model provider would provide
-        its own model events if it wanted to populate the raw request
-        and response fields of the ModelEvent (which only it knows about)
-        """
         return False
 
 
@@ -333,7 +324,7 @@ class Model:
                 )
                 existing = cache_fetch(cache_entry)
                 if isinstance(existing, ModelOutput):
-                    await self._model_transcript(
+                    await self._record_model_interaction(
                         input=input,
                         tools=tools,
                         tool_choice=tool_choice,
@@ -357,7 +348,7 @@ class Model:
                 call = None
 
             # write to transcript
-            await self._model_transcript(
+            await self._record_model_interaction(
                 input=input,
                 tools=tools,
                 tool_choice=tool_choice,
@@ -411,7 +402,7 @@ class Model:
             key=f"Model{self.api.connection_key()}",
         )
 
-    async def _model_transcript(
+    async def _record_model_interaction(
         self,
         input: list[ChatMessage],
         tools: list[ToolInfo],
@@ -423,19 +414,20 @@ class Model:
     ) -> None:
         from inspect_ai.log._transcript import ModelEvent, transcript
 
-        if not self.api.provides_event():
-            transcript()._event(
-                ModelEvent(
-                    model=str(self),
-                    input=input,
-                    tools=tools,
-                    tool_choice=tool_choice,
-                    config=config,
-                    output=output,
-                    cache=cache,
-                    call=call,
-                )
+        trace_assistant_message(input, output.choices[0].message)
+
+        transcript()._event(
+            ModelEvent(
+                model=str(self),
+                input=input,
+                tools=tools,
+                tool_choice=tool_choice,
+                config=config,
+                output=output,
+                cache=cache,
+                call=call,
             )
+        )
 
 
 class ModelName: